refactor(management): split console pages
